UPSTREAM: x86: use _do_fork() in compat_sys_x86_clone()
It is trivial to directly call _do_fork() instead of the sys_clone()
syscall in compat_sys_x86_clone().

This patch is part of a series which removes in-kernel calls to syscalls.
On this basis, the syscall entry path can be streamlined. For details, see
http://lkml.kernel.org/r/20180325162527.GA17492@light.dominikbrodowski.net

Cc: Ingo Molnar <mingo@redhat.com>
Cc: Jiri Slaby <jslaby@suse.com>
Cc: x86@kernel.org
Reviewed-by: Thomas Gleixner <tglx@linutronix.de>
Change-Id: I8fa780e0be29a8b2f184b3e590c58edfadf195f4
Signed-off-by: Dominik Brodowski <linux@dominikbrodowski.net>
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>
(cherry picked from commit 4a62321e7a79440fe5e8631d84e317d7f04f9911)